Solve Day9 and clean up day 8

+++ src/test/kotlin/day7/Day7KtTest.kt
@@ -63,11 +63,11 @@ class Day7KtTest {
 
     }
 
-    private fun testPartB(program: List<Long>, expectedResult: Int) {
+    private fun testPartB(program: List<Long>, expectedResult: Long) {
         assertEquals(expectedResult, solveB(program))
     }
 
-    private fun testPartA(program: List<Int>, expectedResult: Int) {
+    private fun testPartA(program: List<Int>, expectedResult: Long) {
         assertEquals(expectedResult, solveA(program))
     }
